refactoring for 'openPath'
